requirements.txt
django==3.1.2
djangorestframework==3.12.2
django-rest-knox==4.1.0
psycopg2==2.8.6
knox==4.1.0
